+++ messages/de/config.php
@@ -5,7 +5,7 @@
  * NOTE: this file must be saved in UTF-8 encoding.
  */
 return [
-    'AUTH_CONFIG_UPDATE_DESCRIPTION' => 'Einstellungen der Webseite aktualisieren',
+    'AUTH_CONFIG_DESCRIPTION' => 'Website-Einstellungen verwalten',
     'COMMON_SETTINGS' => 'Einstellungen',
     'CONFIG_SUCCESS_UPDATED' => 'Die Einstellungen wurden aktualisiert.',
     'IN_CONFIG' => '',