+++ messages/en-US/config.php
@@ -5,7 +5,7 @@
  * NOTE: this file must be saved in UTF-8 encoding.
  */
 return [
-    'AUTH_CONFIG_UPDATE_DESCRIPTION' => 'Update website settings',
+    'AUTH_CONFIG_DESCRIPTION' => 'Manage website settings',
     'COMMON_SETTINGS' => 'Settings',
     'CONFIG_SUCCESS_UPDATED' => 'The settings were updated.',
     'IN_CONFIG' => 'Config',