+++ messages/pt/config.php
@@ -5,7 +5,7 @@
  * NOTE: this file must be saved in UTF-8 encoding.
  */
 return [
-    'AUTH_CONFIG_UPDATE_DESCRIPTION' => '',
+    'AUTH_CONFIG_DESCRIPTION' => '',
     'COMMON_SETTINGS' => '',
     'CONFIG_SUCCESS_UPDATED' => '',
     'IN_CONFIG' => '',